pyproject.toml
[project]
name = "langgraph-platform-fun-02-client"
version = "0.1.0"
description = ""
authors = [
    {name = "ishish.io",email = "ts@ishish.io"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langgraph-sdk (>=0.1.61,<0.2.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
